requirements.txt
Django==2.2.6
djangorestframework==3.10.3
Faker==2.0.2
mysqlclient==1.4.4
numpy==1.17.2
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
